requirements.txt
opencv-python==4.9.0.80
PyYAML==6.0.1
numpy==1.26.4
tqdm==4.66.2
tabulate==0.9.0